fixing issue with changing ids on address delete

+++ app/controllers/addressController.js
@@ -448,13 +448,22 @@ module.exports.editAddress= function(req,res) {
 };
 
 module.exports.deleteAddress= function(req,res) {
-    Model.SavedAddress.destroy({where:{id:req.query.id}}).then(function() {
-        req.flash('info','Address successfully deleted');
-        return res.redirect('/api/user/addresses');
+    Model.User.findOne({where:{email:req.session.email}}).then(function(user) {
+        Model.SavedAddress.destroy({where: {user_id:user.id, id: req.query.id}}).then(function (result) {
+
+            if(result == true)
+            {
+                console.log("address successfully deleted for user %s and id %s", user.id, req.query.id)
+                req.flash('info', 'Address successfully deleted');
+            }
+            else {
+                console.log("address not deleted for user %s and id %s", user.id, req.query.id)
+                req.flash('info', 'Address not deleted');
+            }
+            return res.redirect('/api/user/addresses');
+        });
     });
-
 };
-
 
 function postcodeLookup(postcode) {
     var rp = require('request-promise');
